+++ api/result.go
@@ -8,11 +8,6 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-// ExecutionResultRequest describes the payload for the REST API request for execution result.
-type ExecutionResultRequest struct {
-	ID string `json:"id"`
-}
-
 // ExecutionResult implements the REST API endpoint for retrieving the result of a function execution.
 func (a *API) ExecutionResult(ctx echo.Context) error {
